requirements.txt
psycopg2-binary
requests
numpy
bs4
PIL
pytest